+++ spec/unit/queue_storage_spec.rb
@@ -45,7 +45,7 @@ module ThreeScale
       def is_sentinel?(connection)
         if ThreeScale::Backend.configuration.redis.async
           connector = connection.instance_variable_get(:@redis_async)
-          connector.instance_of?(Async::Redis::SentinelsClientACLTLS)
+          connector.instance_of?(ThreeScale::Backend::AsyncRedis::SentinelsClientACLTLS)
         else
           connector = connection.instance_variable_get(:@client)
                                 .instance_variable_get(:@config)
